requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
redis>=5.0.1
opencv-python>=4.8.1
pillow>=10.1.0
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.1
pydantic>=2.4.2
numpy>=1.24.0
python-multipart>=0.0.6
